example: Prepare example for CSV collector example
